third_party/WebKit/LayoutTests/inspector/console/console-tainted-globals-expected.txt

Tests that overriding global methods (like Array.prototype.push, Math.max) will not break the inspector.


Running: evaluateInConsole

Running: testRuntimeAgentCallFunctionOn

Running: dumpConsoleMessages
testOverriddenArrayPushAndMathMax()
(3) [1, 2, 3]
testOverriddenConstructorName()
Object {constructor: Object}
testThrowConstructorName()
Object {}
testOverriddenIsFinite()
(2) ["arg1", "arg2", callee: function, Symbol(Symbol.iterator): function]
testOverriddenError()
Object {result: "PASS"}
restoreError()
Object {result: "PASS"}
testOverriddenToString(function func() {}, true)
function #<Function>
testOverriddenToString(function func() {}, false)
function #<Function>
testOverriddenToString(new Function, true)
function #<Function>
testOverriddenToString(new Function, false)
function #<Function>
testOverriddenToString(/^regex$/, true)
#<RegExp>
testOverriddenToString(/^regex$/, false)
#<RegExp>
testOverriddenToString(new Date, true)
#<Date>
testOverriddenToString(new Date, false)
#<Date>
testOverriddenToString({}, true)
Object {toString: function, valueOf: function}
testOverriddenToString({}, false)
Object {toString: function, valueOf: function}
testOverriddenToString(new Number(1), true)
Number {[[PrimitiveValue]]: 1, toString: function, valueOf: function}
testOverriddenToString(new Number(1), false)
Number {[[PrimitiveValue]]: 1, toString: function, valueOf: function}
